autodraw: factor common UI parts
