+++ src/code/css/autodraws.css
@@ -77,16 +77,16 @@
   grid-column: 1 / 4;
 }
 
-.wasabee-dialog-starburst .desc2,
-.wasabee-dialog-fanfield .desc2 {
-  grid-column: 1 / 3;
+.wasabee-dialog-starburst .desc.secondary,
+.wasabee-dialog-fanfield .desc.secondary {
+  color: lightgrey;
 }
 
 .wasabee-dialog-homogeneous .desc {
   grid-column: 1 / 5;
 }
 
-.wasabee-dialog-homogeneous .portal {
+.wasabee-dialog-homogeneous .set-portal-display {
   grid-column: 3 / 5;
 }
 
@@ -98,10 +98,3 @@
   grid-column: 4;
 }
 
-.wasabee-dialog-madrid .set_label {
-  grid-column: 1 / 4;
-}
-
-.wasabee-dialog-starburst .desc2 {
-  color: lightgrey;
-}
