+++ src/code/dialogs/autodraws/starburst.js
@@ -1,20 +1,23 @@
-import { WDialog } from "../../leafletClasses";
+import { AutoDraw } from "./tools";
 import WasabeePortal from "../../model/portal";
 import { getSelectedOperation } from "../../selectedOp";
 import { clearAllLinks, getAllPortalsOnScreen } from "../../uiCommands";
 import wX from "../../wX";
 
-import PortalUI from "../../ui/portal";
-
-const StarburstDialog = WDialog.extend({
+const StarburstDialog = AutoDraw.extend({
   statics: {
     TYPE: "StarburstDialog",
   },
 
-  needWritePermission: true,
+  initialize: function (options) {
+    AutoDraw.prototype.initialize.call(this, options);
+    const p =
+      localStorage[window.plugin.wasabee.static.constants.ANCHOR_ONE_KEY];
+    if (p) this._anchor = new WasabeePortal(p);
+  },
 
   addHooks: function () {
-    WDialog.prototype.addHooks.call(this);
+    AutoDraw.prototype.addHooks.call(this);
     this._displayDialog();
   },
 
@@ -24,40 +27,14 @@ const StarburstDialog = WDialog.extend({
     const description = L.DomUtil.create("div", "desc", container);
     description.textContent = wX("SEL_SB_ANCHOR");
 
-    //anchor portal text
-    const anchorLabel = L.DomUtil.create("label", null, container);
-    anchorLabel.textContent = wX("ANCHOR_PORTAL");
+    this._addSetPortal(
+      wX("ANCHOR_PORTAL"),
+      "_anchor",
+      container,
+      window.plugin.wasabee.static.constants.ANCHOR_ONE_KEY
+    );
 
-    //Set Button
-    const anchorButton = L.DomUtil.create("button", null, container);
-    anchorButton.textContent = wX("SET");
-    this._anchorDisplay = L.DomUtil.create("div", "anchor", container);
-
-    //do magic
-    if (this._anchor) {
-      this._anchorDisplay.appendChild(
-        PortalUI.displayFormat(this._anchor, this._smallScreen)
-      );
-    } else {
-      this._anchorDisplay.textContent = wX("NOT_SET");
-    }
-
-    L.DomEvent.on(anchorButton, "click", (ev) => {
-      L.DomEvent.stop(ev);
-      this._anchor = PortalUI.getSelected();
-      if (this._anchor) {
-        localStorage[window.plugin.wasabee.static.constants.ANCHOR_ONE_KEY] =
-          JSON.stringify(this._anchor);
-        this._anchorDisplay.textContent = "";
-        this._anchorDisplay.appendChild(
-          PortalUI.displayFormat(this._anchor, this._smallScreen)
-        );
-      } else {
-        alert(wX("PLEASE_SELECT_PORTAL"));
-      }
-    });
-
-    const description2 = L.DomUtil.create("div", "desc2", container);
+    const description2 = L.DomUtil.create("div", "desc secondary", container);
     description2.textContent = wX("SEL_SB_ANCHOR2");
 
     const button = L.DomUtil.create("button", "drawb", container);
@@ -83,13 +60,6 @@ const StarburstDialog = WDialog.extend({
       buttons: buttons,
       id: window.plugin.wasabee.static.dialogNames.starburst,
     });
-  },
-
-  initialize: function (options) {
-    WDialog.prototype.initialize.call(this, options);
-    const p =
-      localStorage[window.plugin.wasabee.static.constants.ANCHOR_ONE_KEY];
-    if (p) this._anchor = new WasabeePortal(p);
   },
 
   starburst: function () {
